requirements.txt
pdfminer.six==20221105
scikit-learn==1.4.2
numpy==1.26.4